Update CruncherSharp with alignment (computed, not 100% accurate) Hardcode though config some alignment (vectors and quats) Add potentail saving, calculated from paddingm, alignment, unused vtables... Update columns for readibility #rnx [FYI] Marc.Audy
[CL 35319747 by charles lefebvre in ue5-main branch]

+++ Engine/Extras/ThirdPartyNotUE/crunchersharp/SymbolAnalyzerRawPDB.cs
@@ -285,7 +285,7 @@ namespace CruncherSharp
 					typeName,
 					typeSymbol.Size, 
 					dataMember.BitSize, 
-					(ulong)dataMember.Offset,
+					dataMember.Offset,
 					dataMember.BitPosition 
 				);
 				info.BitField = dataMember.IsBitfield;
